agregar y quitar productos del carrito

+++ src/data/cartService.js
@@ -8,7 +8,7 @@ const cartService = {
             let carritoNuevo = await db.Carritos.create({
                 ID_Usuario: idUser,
                 Fecha_Creacion: new Date(),
-                Total: 0
+                Total: '0.00'
             });
             return carritoNuevo;
         } catch (error) {
@@ -55,20 +55,80 @@ const cartService = {
             return [];
         }
     },
-
+    getDetalle: async function (idDetalle) {
+        try {
+            let detalle = await db.DetalleCarrito.findOne({
+                where: {
+                    id: idDetalle
+                }, raw: true
+            });
+            return detalle;
+        } catch (error) {
+            console.log("Error al crear el carrito", error);
+            return [];
+        }
+    },
+    getBy: async function (whereClausure = {}) {
+        return await db.Carritos.findAll({ where: whereClausure });
+    },
+    existeProductoEnCarrito: async function (idCar, idProd) {
+        return await db.DetalleCarrito.findOne({
+            where: {
+                ID_Carrito: idCar,
+                ID_Producto: idProd
+            }
+        })
+    },
+    calcularTotalCarritoIncremento: function (totalCarrito, precioProdu, descProdu, nuevaCantidad) {
+        totalCarrito = parseFloat(totalCarrito);
+        let precioConDescuento = parseFloat(precioProdu) - (parseFloat(precioProdu) * (descProdu / 100));
+        let totalDetalle = nuevaCantidad * precioConDescuento;
+        let nuevoTotalCarrito = totalCarrito + totalDetalle;
+        return nuevoTotalCarrito
+    },
+    calcularTotalCarritoDecremento: function (totalCarrito, precioProdu, descProdu, nuevaCantidad) {
+        totalCarrito = parseFloat(totalCarrito);
+        let precioConDescuento = parseFloat(precioProdu) - (parseFloat(precioProdu) * (descProdu / 100));
+        let totalDetalle = nuevaCantidad * precioConDescuento;
+        let nuevoTotalCarrito = totalCarrito - totalDetalle;
+        return nuevoTotalCarrito
+    },
     addProductoAlCarrito: async function (idCarrito, idProducto, cantidad) {
         try {
-            let nuevoDetalleCarrito = await db.DetalleCarrito.create({
-                ID_Carrito: idCarrito,
-                ID_Producto: idProducto,
-                Cantidad: cantidad
-            });
-            return nuevoDetalleCarrito;
+            let nuevoTotalCarrito;
+            let producto = await db.Productos.findByPk(idProducto);
+            if (!producto) {
+                throw new Error('No se encontró el producto');
+            }
+            const existeProducto = await this.existeProductoEnCarrito(idCarrito, idProducto);
+            const existeCarrito = await this.getBy({ id: idCarrito });
+            if (existeProducto) {
+                await db.DetalleCarrito.update(
+                    { Cantidad: parseInt(cantidad) + existeProducto.Cantidad },
+                    {
+                        where: {
+                            ID_Producto: idProducto,
+                            ID_Carrito: idCarrito
+                        }
+                    }
+                );
+            } else {
+                await db.DetalleCarrito.create({
+                    ID_Carrito: idCarrito,
+                    ID_Producto: idProducto,
+                    Cantidad: cantidad
+                });
+            }
+            productoDetalle = await db.DetalleCarrito.findOne({ where: { ID_Carrito: idCarrito, ID_Producto: idProducto } });
+            nuevoTotalCarrito = this.calcularTotalCarritoIncremento(existeCarrito[0].Total, producto.Precio, producto.Descuento, productoDetalle.Cantidad)
+
+            return await db.Carritos.update({ Total: nuevoTotalCarrito }, { where: { id: idCarrito } });
         } catch (error) {
             console.error('Error al agregar producto al carrito:', error);
             throw error;
         }
     },
+    //TODO: Cuandop quito detalles eliminar carrito si queda vacio
     getProductosCarrito: async function (idCart) {
         try {
             const productosCarrito = await db.DetalleCarrito.findAll({
@@ -89,41 +149,78 @@ const cartService = {
             return [];
         }
     },
-    getCantidadByIdProdIdCart: async function (idCarrito,idProd) {
+    getCantidadByIdProdIdCart: async function (idCarrito, idProd) {
         try {
             const cantidadProducto = await db.DetalleCarrito.findOne({
-                where: { ID_Producto: idProd,ID_Carrito:idCarrito },
+                where: { ID_Producto: idProd, ID_Carrito: idCarrito },
                 raw: true
             });
             return cantidadProducto;
         } catch (error) {
             console.log(error);
-            return[];
+            return [];
         }
     },
-    updateCantidad: async function (idDetalleProducto, cantidad) {
+    updateCantidad: async function (idDetalleProducto, nuevaCantidad) {
         try {
-            const detalleProducto = await db.DetalleCarrito.findByPk(idDetalleProducto);
-            if (!detalleProducto) {
-                throw new Error('Elemento del carrito no encontrado');
+            let detalleProductoAnterior = await db.DetalleCarrito.findByPk(idDetalleProducto);
+
+            await db.DetalleCarrito.update({ Cantidad: nuevaCantidad }, { where: { id: idDetalleProducto } });
+
+            let detalleProductoActualizado = await db.DetalleCarrito.findByPk(idDetalleProducto);
+
+            let diferenciaCantidad = nuevaCantidad - detalleProductoAnterior.Cantidad;
+
+            let producto = await db.Productos.findByPk(detalleProductoActualizado.ID_Producto);
+
+            let carrito = await db.Carritos.findByPk(detalleProductoActualizado.ID_Carrito);
+
+            let nuevoTotalCarrito = carrito.Total;
+
+            if (diferenciaCantidad < 0) { // Si estás decrementando la cantidad
+                nuevoTotalCarrito = this.calcularTotalCarritoDecremento(carrito.Total, producto.Precio, producto.Descuento, Math.abs(diferenciaCantidad))
+            } else if (diferenciaCantidad > 0) { // Si estás incrementando la cantidad
+                nuevoTotalCarrito = this.calcularTotalCarritoIncremento(carrito.Total, producto.Precio, producto.Descuento, Math.abs(diferenciaCantidad))
             }
-            detalleProducto.Cantidad = cantidad;
-            await this.actualizarTotalCarrito(detalleProducto.ID_Carrito, detalleProducto.ID_Producto, cantidad);
-            await detalleProducto.save();
-            return detalleProducto;
+
+            // Actualizar el total del carrito en la base de datos
+            return await db.Carritos.update({ Total: nuevoTotalCarrito }, { where: { id: carrito.id } });
+
         } catch (error) {
             console.log(error);
             return [];
         }
     },
+
     delete: async function (id) {
         try {
-            return await db.DetalleCarrito.destroy({ where: { id: id } });
+            let detalleEliminar = await this.getDetalle(id);
+            if (!detalleEliminar) {
+                throw new Error('No se encontró el detalle del producto a eliminar');
+            }
+
+            let productoEliminar = await db.Productos.findByPk(detalleEliminar.ID_Producto);
+            if (!productoEliminar) {
+                throw new Error('No se encontró el producto a eliminar');
+            }
+            await db.DetalleCarrito.destroy({ where: { id: id } });
+
+            let carrito = await db.Carritos.findByPk(detalleEliminar.ID_Carrito);
+            if (!carrito) {
+                throw new Error('No se encontró el carrito asociado al detalle del producto eliminado');
+            }
+
+            let nuevoTotalCarrito = this.calcularTotalCarritoDecremento(carrito.Total, productoEliminar.Precio, productoEliminar.Descuento, detalleEliminar.Cantidad);
+
+            await db.Carritos.update({ Total: nuevoTotalCarrito }, { where: { id: carrito.id } });
+
+            return true;
         } catch (error) {
             console.log(error);
-            return [];
+            return false;
         }
     }
+
 
 }
 module.exports = cartService;
